add authentication and jwt

+++ src/resume/resume.controller.ts
@@ -1,24 +1,35 @@
 // resume/resume.controller.ts
-import { Controller, Get, Post, Body, Param, Put, Delete, Query } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Post,
+  Body,
+  Param,
+  Put,
+  Delete,
+  Query,
+  UseGuards,
+  Req,
+  NotFoundException,
+  ForbiddenException,
+} from '@nestjs/common';
 import { ResumeService } from './resume.service';
 import { CreateResumeDto } from './dto/create-resume.dto';
 import { UpdateResumeDto } from './dto/update-resume.dto';
 import { Resume } from './entities/resume.entity';
-import { ApiTags } from '@nestjs/swagger';
+import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
 import { AgeCriteriaDto } from './dto/age-criteria.dto';
+import { AuthMiddleware } from 'src/middlewares/auth.middleware';
+import { AuthGuard } from '@nestjs/passport';
 
 @ApiTags('resumes')
 @Controller('resumes')
+@UseGuards(AuthMiddleware)
 export class ResumeController {
   constructor(private readonly resumeService: ResumeService) {}
 
-  @Post()
-  async createResume(@Body() createResumeDto: CreateResumeDto): Promise<Resume> {
-    return this.resumeService.createResume(createResumeDto);
-  }
-
   @Get()
-  async getResumes(@Query() ageCriteriaDto: AgeCriteriaDto) : Promise<Resume[]> {
+  async getResumes(@Query() ageCriteriaDto: AgeCriteriaDto): Promise<Resume[]> {
     return this.resumeService.getResumes(ageCriteriaDto);
   }
 
@@ -27,13 +38,53 @@ export class ResumeController {
     return this.resumeService.getResumeById(id);
   }
 
+  @Post()
+  @UseGuards(AuthGuard('jwt'))
+  @ApiBearerAuth()
+  async createResume(
+    @Body() createResumeDto: CreateResumeDto,
+  ): Promise<Resume> {
+    return this.resumeService.createResume(createResumeDto);
+  }
+
   @Put(':id')
-  async updateResume(@Param('id') id: string, @Body() updateResumeDto: UpdateResumeDto): Promise<Resume> {
+  @UseGuards(AuthGuard('jwt'))
+  @ApiBearerAuth()
+  async updateResume(
+    @Param('id') id: string,
+    @Body() updateResumeDto: UpdateResumeDto,
+    @Req() req: Request,
+  ): Promise<Resume> {
+    const userId = req['id']; // Get the user id from the request
+    const resume = await this.resumeService.getResumeById(id);
+
+    if (!resume) {
+      throw new NotFoundException('Resume not found');
+    }
+
+    if (userId !== resume.user.id) {
+      throw new ForbiddenException('You are not allowed to update this resume');
+    }
     return this.resumeService.updateResume(id, updateResumeDto);
   }
 
   @Delete(':id')
-  async deleteResume(@Param('id') id: string): Promise<void> {
+  @UseGuards(AuthGuard('jwt'))
+  @ApiBearerAuth()
+  async deleteResume(
+    @Param('id') id: string,
+    @Req() req: Request,
+  ): Promise<void> {
+    const userId = req['id']; // Get the user id from the request
+    const resume = await this.resumeService.getResumeById(id);
+
+    if (!resume) {
+      throw new NotFoundException('Resume not found');
+    }
+
+    if (userId !== resume.user.id) {
+      throw new ForbiddenException('You are not allowed to delete this resume');
+    }
     return this.resumeService.deleteResume(id);
   }
 }
