+++ src/user/user.controller.ts
@@ -1,20 +1,35 @@
 // user/user.controller.ts
-import { Controller, Get, Post, Body, Param, Put, Delete } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Post,
+  Body,
+  Param,
+  Put,
+  Delete,
+  UseGuards,
+  UnauthorizedException,
+  Req,
+  NotFoundException,
+  ForbiddenException,
+} from '@nestjs/common';
 import { UserService } from './user.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { User } from './entities/user.entity';
-import { ApiTags } from '@nestjs/swagger';
+import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
+import { AuthMiddleware } from 'src/middlewares/auth.middleware';
+import { LoginDto } from './dto/login.dto';
+import { AuthService } from './auth.service';
+import { AuthGuard } from '@nestjs/passport';
 
 @ApiTags('users')
 @Controller('users')
 export class UserController {
-  constructor(private readonly userService: UserService) {}
-
-  @Post()
-  async createUser(@Body() createUserDto: CreateUserDto): Promise<User> {
-    return this.userService.createUser(createUserDto);
-  }
+  constructor(
+    private readonly userService: UserService,
+    private readonly authService: AuthService,
+  ) {}
 
   @Get()
   async getUsers(): Promise<User[]> {
@@ -26,13 +41,64 @@ export class UserController {
     return this.userService.getUserById(id);
   }
 
+  @Post()
+  async createUser(@Body() createUserDto: CreateUserDto): Promise<User> {
+    return this.userService.createUser(createUserDto);
+  }
+
   @Put(':id')
-  async updateUser(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto): Promise<User> {
+  @UseGuards(AuthGuard('jwt'))
+  @ApiBearerAuth()
+  async updateUser(
+    @Param('id') id: string,
+    @Body() updateUserDto: UpdateUserDto,
+    @Req() req: Request,
+  ): Promise<User> {
+    const userId = req['userId']; // Get the user id from the request
+    const user = await this.userService.getUserById(id);
+
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+
+    if (userId !== user.id) {
+      console.log(userId +"    " +user.id)
+      throw new ForbiddenException('You are not allowed to update this user');
+    }
+
     return this.userService.updateUser(id, updateUserDto);
   }
 
   @Delete(':id')
-  async deleteUser(@Param('id') id: string): Promise<void> {
-    return this.userService.deleteUser(id);
+  @UseGuards(AuthGuard('jwt'), AuthMiddleware)
+  @ApiBearerAuth()
+  async deleteUser(
+    @Param('id') id: string,
+    @Req() req: Request,
+  ): Promise<void> {
+    const userId = req['id']; // Get the user id from the request
+    const user = await this.userService.getUserById(id);
+
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+
+    if (userId !== user.id) {
+      throw new ForbiddenException('You are not allowed to delete this user');
+    }
+
+    await this.userService.deleteUser(id);
+  }
+
+  @Post('login')
+  async login(@Body() loginDto: LoginDto) {
+    const user = await this.authService.validateUser(
+      loginDto.email,
+      loginDto.password,
+    );
+    if (!user) {
+      throw new UnauthorizedException();
+    }
+    return this.authService.login(user);
   }
 }
